requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
clip==0.2.0
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lightning-utilities==0.14.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.4.2
nudenet==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scalene==1.5.51
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchmetrics==1.7.4
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
yarl==1.20.1
